Move Mapbox related code to a module
